Add verification check for more per-frame sequence items than frames in a DICOM file.

+++ zz_priv.h
@@ -152,7 +152,7 @@ struct zzfile
 		uint16_t	group;
 		uint16_t	element;
 		long		pos;		// start of a current tag's data segment
-		int		frame;		// current frame number, or -1, zero indexed
+		long		frame;		// current frame number, or -1, zero indexed
 		char		warning[MAX_LEN_LO];	// if !valid, this string is set to an explanation
 		bool		valid;		// whether the current tag is valid, true unless issue found
 		enum zzpxstate	pxstate;
